# Speed up XZ decompression by 5x with bufio wrapper