--- a/decompress_txz.go
+++ b/decompress_txz.go
@@ -4,6 +4,7 @@
 package getter
 
 import (
+	"bufio"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -45,7 +46,7 @@ func (d *TarXzDecompressor) Decompress(dst, src string, dir bool, umask os.FileM
 	defer f.Close()
 
 	// xz compression is second
-	txzR, err := xz.NewReader(f)
+	txzR, err := xz.NewReader(bufio.NewReader(f))
 	if err != nil {
 		return fmt.Errorf("Error opening an xz reader for %s: %s", src, err)
 	}