--- a/decompress_xz.go
+++ b/decompress_xz.go
@@ -4,6 +4,7 @@
 package getter
 
 import (
+	"bufio"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -39,7 +40,7 @@ func (d *XzDecompressor) Decompress(dst, src string, dir bool, umask os.FileMode
 	defer f.Close()
 
 	// xz compression is second
-	xzR, err := xz.NewReader(f)
+	xzR, err := xz.NewReader(bufio.NewReader(f))
 	if err != nil {
 		return err
 	}