Nomad was taking over 7.5 seconds just to extract an 11MB tar.xz archive for our use case.
Investigation led to XZ library which performs poorly without buffered I/O (https://github.com/ulikunitz/xz/issues/23).

Decompression performance can be improved dramatically by introducing buffered I/O as suggested in the case.

Benchmark results (using hyperfine with identical test files served from a local `python3 -m http.server 8888` web server):

Before:
```
./go-getter http://localhost:8888/test.tar.xz tmp-xz
Time (mean ± σ): 7.721s ± 0.019s [User: 3.541s, System: 4.225s]
Range (min … max): 7.685s … 7.749s (10 runs)
```

After:
```
./go-getter http://localhost:8888/test.tar.xz tmp-xz
Time (mean ± σ): 1.429s ± 0.025s [User: 1.378s, System: 0.061s]
Range (min … max): 1.411s … 1.495s (10 runs)
```

This represents a 5.4x speedup and should significantly improve performance when working with XZ-compressed artifacts.