@crw 2025-03-18
Hi @vsarunas, are you making this change to target a specific HashiCorp product? Terraform is using the 1.x tag, and Packer is using the 2.x tag, so this change will likely need to be backported to one or the other (I suspect backported to 1.x but not entirely sure which tag `main` builds natively.)