@vsarunas 2025-03-18
Hi @crw, yes interested in Nomad which uses currently v1.7.8. I do not see a v1 specific branch. I should branch from the tag v1.7.8 release and have 2 PRs?