@dduzgun-security 2025-08-15
@vsarunas a new release is now available [v1.7.9](https://github.com/hashicorp/go-getter/releases/tag/v1.7.9), thanks for the contribution. 🙏